revert go mod

+++ br/pkg/lightning/backend/local/local_test.go
@@ -279,7 +279,7 @@ func TestRangePropertiesWithPebble(t *testing.T) {
 			binary.BigEndian.PutUint64(key, uint64(i*100+j))
 			err = wb.Set(key, value[:valueLen], writeOpt)
 			require.NoError(t, err)
-			err = collector.Add(pebble.InternalKey{UserKey: key, Trailer: uint64(pebble.InternalKeyKindSet)}, value[:valueLen])
+			err = collector.Add(pebble.InternalKey{UserKey: key, Trailer: pebble.InternalKeyKindSet}, value[:valueLen])
 			require.NoError(t, err)
 		}
 		require.NoError(t, wb.Commit(writeOpt))
